requirements.txt
tensorboardX
tqdm
easydict
prettytable
chumpy
opencv-python
imageio-ffmpeg
matplotlib==3.7.5
roma
ipdb
pytorch-metric-learning                         # For one-shot action recognition
smplx[all]                                      # For mesh recovery
mediapipe==0.10.11